Final adjustments to Logic
